requirements.txt
# FastAPI and WebSocket dependencies
fastapi>=0.109.0
websockets>=12.0
uvicorn>=0.27.0
python-dotenv>=1.0.0

# LangChain and its dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.16

# Vector store and embeddings
chromadb>=0.4.22
langchain-chroma>=0.0.12

# Google AI dependencies
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.1

# CORS middleware
starlette>=0.36.0

# Type hints support
typing-extensions>=4.9.0

# JSON handling
pyjson5>=1.6.5

# Optional: Performance improvements
httpx>=0.26.0